requirements.txt
streamlit
pandas
openpyxl
python-docx
fpdf2
qrcode
openai
requests
Pillow